refactor(mitm-addon): split websocket tests by responsibility
