+++ crates/runner/mitm-addon/tests/test_model_provider_websocket_usage.py
@@ -6,8 +6,6 @@ from pathlib import Path
 
 import pytest
 from mitmproxy import http
-from mitmproxy.flow import Error
-from mitmproxy.test import tutils
 
 import flow_metadata_keys as metadata_keys
 import mitm_addon
@@ -15,22 +13,16 @@ import usage
 from tests.jsonl_log_helpers import jsonl_exists_after_flush, read_jsonl_entries_after_flush
 from tests.model_provider_flow_helpers import (
     make_openai_responses_websocket_flow,
-    make_openai_responses_websocket_request_flow,
-    make_openai_responses_websocket_response_headers,
     model_provider_usage_sources,
 )
 from tests.model_provider_websocket_helpers import (
     ScheduledWebSocketTrim,
-    append_websocket_message,
     capture_deferred_websocket_trims,
     feed_websocket_server_message,
     feed_websocket_server_text_message,
     openai_websocket_usage_frame,
-    run_deferred_websocket_trims,
     set_websocket_message,
 )
-from tests.pending_helpers import assert_pending
-from tests.request_handler_helpers import _single_firewall_vm, _write_registry
 
 
 @pytest.fixture(autouse=True)
@@ -38,35 +30,6 @@ def deferred_websocket_trim_scheduler(
     monkeypatch: pytest.MonkeyPatch,
 ) -> list[ScheduledWebSocketTrim]:
     return capture_deferred_websocket_trims(monkeypatch)
-
-
-def _write_openai_model_websocket_registry(tmp_path: Path) -> Path:
-    firewall_name = "model-provider:openai-api-key"
-    return _write_registry(
-        tmp_path,
-        vm_info=_single_firewall_vm(
-            tmp_path,
-            run_id="run-abc-123",
-            sandbox_marker="tok-xyz",
-            firewall_name=firewall_name,
-            api_entry={
-                "base": "https://api.openai.com",
-                "auth": {"headers": {"Authorization": "Bearer token"}},
-                "permissions": [{"name": "responses", "rules": ["GET /v1/responses"]}],
-            },
-            network_policy={
-                "allow": ["responses"],
-                "deny": [],
-                "ask": [],
-                "unknownPolicy": "deny",
-            },
-            billable_firewalls=[firewall_name],
-            vm_fields={
-                "cliAgentType": "codex",
-                "modelUsageProvider": "gpt-5.5",
-            },
-        ),
-    )
 
 
 def _sum_quantities_by_category(events: list[dict]) -> dict[str, int]:
@@ -102,91 +65,8 @@ def _openai_websocket_zero_usage_frame(response_id: str, *, model: str | None = 
     ).encode()
 
 
-class TestModelProviderWebSocketUsage:
-    """Tests for model-provider WebSocket usage reporting."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
-        self._usage_webhook_api = usage_webhook_api
-
-    def _run_response(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.response(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_error(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_websocket_end(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_websocket_messages_and_end(self, flow: http.HTTPFlow, *messages: bytes):
-        with self._usage_webhook_api() as webhook:
-            for message in messages:
-                feed_websocket_server_message(flow, message)
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_websocket_message_and_end(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def test_full_pipeline_model_websocket_reports_usage(self, tmp_path, real_flow):
-        """Codex Responses WebSocket frames should bill like SSE events."""
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        assert flow.metadata["model_websocket_usage_enabled"] is True
-        assert "model_json_usage_finish" not in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-        set_websocket_message(
-            flow,
-            from_client=False,
-            content=json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {
-                            "input_tokens": 50,
-                            "output_tokens": 20,
-                            "input_tokens_details": {
-                                "cached_tokens": 10,
-                                "cache_write_tokens": 15,
-                            },
-                        },
-                    },
-                }
-            ).encode(),
-        )
-
-        webhook = self._run_websocket_message_and_end(flow)
-
-        events = webhook.usage_events()
-        by_category = _sum_quantities_by_category(events)
-        assert len(events) == len(by_category)
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert model_provider_usage_sources(flow) == {}
-        assert by_category == {
-            "tokens.input": 25,
-            "tokens.output": 20,
-            "tokens.cache_read": 10,
-            "tokens.cache_creation": 15,
-        }
+class TestModelProviderWebSocketUsageSourceRelease:
+    """Tests for sources that cannot be delivered to the usage webhook."""
 
     def test_model_websocket_missing_context_releases_positive_source(
         self, tmp_path, real_flow, mitm_ctx
@@ -289,6 +169,75 @@ class TestModelProviderWebSocketUsage:
 
         assert model_provider_usage_sources(flow) == {}
         assert not jsonl_exists_after_flush(proxy_log)
+
+
+class TestModelProviderWebSocketUsage:
+    """Tests for model-provider WebSocket usage reporting."""
+
+    @pytest.fixture(autouse=True)
+    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
+        self._usage_webhook_api = usage_webhook_api
+
+    def _run_websocket_messages_and_end(self, flow: http.HTTPFlow, *messages: bytes):
+        with self._usage_webhook_api() as webhook:
+            for message in messages:
+                feed_websocket_server_message(flow, message)
+            mitm_addon.websocket_end(flow)
+            usage.flush_usage_events(trigger="test")
+        return webhook
+
+    def _run_websocket_message_and_end(self, flow: http.HTTPFlow):
+        with self._usage_webhook_api() as webhook:
+            mitm_addon.websocket_message(flow)
+            mitm_addon.websocket_end(flow)
+            usage.flush_usage_events(trigger="test")
+        return webhook
+
+    def test_full_pipeline_model_websocket_reports_usage(self, tmp_path, real_flow):
+        """Codex Responses WebSocket frames should bill like SSE events."""
+        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
+        mitm_addon.responseheaders(flow)
+        assert flow.metadata["model_websocket_usage_enabled"] is True
+        assert "model_json_usage_finish" not in flow.metadata
+        assert "model_sse_usage_finish" not in flow.metadata
+        assert metadata_keys.STREAM_BUFFER not in flow.metadata
+        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
+
+        set_websocket_message(
+            flow,
+            from_client=False,
+            content=json.dumps(
+                {
+                    "type": "response.completed",
+                    "response": {
+                        "id": "resp_ws_1",
+                        "model": "gpt-5.5",
+                        "usage": {
+                            "input_tokens": 50,
+                            "output_tokens": 20,
+                            "input_tokens_details": {
+                                "cached_tokens": 10,
+                                "cache_write_tokens": 15,
+                            },
+                        },
+                    },
+                }
+            ).encode(),
+        )
+
+        webhook = self._run_websocket_message_and_end(flow)
+
+        events = webhook.usage_events()
+        by_category = _sum_quantities_by_category(events)
+        assert len(events) == len(by_category)
+        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
+        assert model_provider_usage_sources(flow) == {}
+        assert by_category == {
+            "tokens.input": 25,
+            "tokens.output": 20,
+            "tokens.cache_read": 10,
+            "tokens.cache_creation": 15,
+        }
 
     def test_full_pipeline_model_websocket_reports_multiple_response_ids(self, tmp_path, real_flow):
         flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
@@ -833,277 +782,6 @@ class TestModelProviderWebSocketUsage:
         assert observations_by_category == by_category
         assert model_provider_usage_sources(flow) == {}
 
-    async def test_model_websocket_response_keeps_usage_flow_tracked_until_end(
-        self,
-        tmp_path,
-        real_flow,
-        mitm_ctx,
-        fake_firewall_headers,
-        usage_webhook_server,
-    ):
-        """The HTTP 101 response hook must not complete the WebSocket usage lifecycle."""
-        pending_path = tmp_path / "usage-pending"
-        usage.set_pending_path(str(pending_path), usage_state_id="test-usage-state-id")
-        reg_path = _write_openai_model_websocket_registry(tmp_path)
-
-        flow = make_openai_responses_websocket_request_flow(real_flow)
-
-        with (
-            mitm_ctx(registry_path=str(reg_path), api_url=usage_webhook_server.api_url),
-            fake_firewall_headers(),
-        ):
-            await mitm_addon.request(flow)
-            usage.write_pending_snapshot(flush_request_id="before-response")
-            assert_pending(
-                pending_path,
-                flows=1,
-                buffered=0,
-                reports=0,
-                flush_request_id="before-response",
-            )
-
-            flow.response = tutils.tresp(
-                status_code=101,
-                headers=make_openai_responses_websocket_response_headers(),
-            )
-            mitm_addon.responseheaders(flow)
-            mitm_addon.response(flow)
-            usage.write_pending_snapshot(flush_request_id="after-response")
-            assert_pending(
-                pending_path,
-                flows=1,
-                buffered=0,
-                reports=0,
-                flush_request_id="after-response",
-            )
-
-            feed_websocket_server_message(
-                flow,
-                json.dumps(
-                    {
-                        "type": "response.completed",
-                        "response": {
-                            "id": "resp_ws_1",
-                            "model": "gpt-5.5",
-                            "usage": {
-                                "input_tokens": 50,
-                                "output_tokens": 20,
-                                "input_tokens_details": {"cached_tokens": 10},
-                            },
-                        },
-                    }
-                ).encode(),
-            )
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-
-        events = usage_webhook_server.usage_events()
-        assert len(events) == 3
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {
-            "tokens.input": 40,
-            "tokens.output": 20,
-            "tokens.cache_read": 10,
-        }
-        observation_events = usage_webhook_server.model_usage_observation_events()
-        assert len(observation_events) == len(events)
-        assert {event["category"]: event["quantity"] for event in observation_events} == by_category
-        assert {event["model"] for event in observation_events} == {"gpt-5.5"}
-        usage.write_pending_snapshot(flush_request_id="after-websocket-end")
-        assert_pending(
-            pending_path,
-            flows=0,
-            buffered=0,
-            reports=0,
-            flush_request_id="after-websocket-end",
-        )
-
-    async def test_non_websocket_switching_protocols_response_releases_usage_flow(
-        self,
-        tmp_path,
-        real_flow,
-        mitm_ctx,
-        fake_firewall_headers,
-        usage_webhook_server,
-    ):
-        """A non-WebSocket 101 response is terminal and must not wait for websocket_end()."""
-        pending_path = tmp_path / "usage-pending"
-        usage.set_pending_path(str(pending_path), usage_state_id="test-usage-state-id")
-        reg_path = _write_openai_model_websocket_registry(tmp_path)
-
-        flow = real_flow(
-            with_response=False,
-            client_ip="10.200.0.5",
-            host="api.openai.com",
-            path="/v1/responses",
-            method="GET",
-        )
-
-        with (
-            mitm_ctx(registry_path=str(reg_path), api_url=usage_webhook_server.api_url),
-            fake_firewall_headers(),
-        ):
-            await mitm_addon.request(flow)
-            usage.write_pending_snapshot(flush_request_id="before-response")
-            assert_pending(
-                pending_path,
-                flows=1,
-                buffered=0,
-                reports=0,
-                flush_request_id="before-response",
-            )
-
-            flow.response = tutils.tresp(
-                status_code=101,
-                headers=http.Headers(upgrade="h2c"),
-            )
-            mitm_addon.responseheaders(flow)
-            mitm_addon.response(flow)
-            usage.write_pending_snapshot(flush_request_id="after-response")
-
-        assert_pending(
-            pending_path,
-            flows=0,
-            buffered=0,
-            reports=0,
-            flush_request_id="after-response",
-        )
-
-    @pytest.mark.parametrize(
-        "response_headers",
-        [
-            pytest.param(
-                make_openai_responses_websocket_response_headers(upgrade="h2c"),
-                id="non-websocket-upgrade",
-            ),
-            pytest.param(
-                make_openai_responses_websocket_response_headers(connection=None),
-                id="missing-connection-upgrade",
-            ),
-            pytest.param(
-                make_openai_responses_websocket_response_headers(accept="wrong"),
-                id="wrong-accept",
-            ),
-        ],
-    )
-    async def test_invalid_websocket_switching_protocols_response_releases_usage_flow(
-        self,
-        tmp_path,
-        real_flow,
-        mitm_ctx,
-        fake_firewall_headers,
-        usage_webhook_server,
-        response_headers: http.Headers,
-    ):
-        """A malformed 101 WebSocket response must not wait for websocket_end()."""
-        pending_path = tmp_path / "usage-pending"
-        usage.set_pending_path(str(pending_path), usage_state_id="test-usage-state-id")
-        reg_path = _write_openai_model_websocket_registry(tmp_path)
-
-        flow = make_openai_responses_websocket_request_flow(real_flow)
-
-        with (
-            mitm_ctx(registry_path=str(reg_path), api_url=usage_webhook_server.api_url),
-            fake_firewall_headers(),
-        ):
-            await mitm_addon.request(flow)
-            usage.write_pending_snapshot(flush_request_id="before-response")
-            assert_pending(
-                pending_path,
-                flows=1,
-                buffered=0,
-                reports=0,
-                flush_request_id="before-response",
-            )
-
-            flow.response = tutils.tresp(status_code=101, headers=response_headers)
-            mitm_addon.responseheaders(flow)
-            mitm_addon.response(flow)
-            usage.write_pending_snapshot(flush_request_id="after-response")
-
-        assert_pending(
-            pending_path,
-            flows=0,
-            buffered=0,
-            reports=0,
-            flush_request_id="after-response",
-        )
-
-    async def test_model_websocket_error_releases_usage_flow_after_upgrade(
-        self,
-        tmp_path,
-        real_flow,
-        mitm_ctx,
-        fake_firewall_headers,
-        usage_webhook_server,
-    ):
-        """A WebSocket connection error after HTTP 101 is terminal for usage tracking."""
-        pending_path = tmp_path / "usage-pending"
-        usage.set_pending_path(str(pending_path), usage_state_id="test-usage-state-id")
-        reg_path = _write_openai_model_websocket_registry(tmp_path)
-
-        flow = make_openai_responses_websocket_request_flow(real_flow)
-
-        with (
-            mitm_ctx(registry_path=str(reg_path), api_url=usage_webhook_server.api_url),
-            fake_firewall_headers(),
-        ):
-            await mitm_addon.request(flow)
-            flow.response = tutils.tresp(
-                status_code=101,
-                headers=make_openai_responses_websocket_response_headers(),
-            )
-            mitm_addon.responseheaders(flow)
-            mitm_addon.response(flow)
-            usage.write_pending_snapshot(flush_request_id="after-response")
-            assert_pending(
-                pending_path,
-                flows=1,
-                buffered=0,
-                reports=0,
-                flush_request_id="after-response",
-            )
-
-            feed_websocket_server_message(
-                flow,
-                json.dumps(
-                    {
-                        "type": "response.completed",
-                        "response": {
-                            "id": "resp_ws_error",
-                            "model": "gpt-5.5",
-                            "usage": {
-                                "input_tokens": 10,
-                                "output_tokens": 4,
-                            },
-                        },
-                    }
-                ).encode(),
-            )
-            flow.error = Error("connection reset by peer")
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-
-        events = usage_webhook_server.usage_events()
-        assert len(events) == 2
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert by_category == {
-            "tokens.input": 10,
-            "tokens.output": 4,
-        }
-        observation_events = usage_webhook_server.model_usage_observation_events()
-        assert len(observation_events) == len(events)
-        assert {event["category"]: event["quantity"] for event in observation_events} == by_category
-        assert {event["model"] for event in observation_events} == {"gpt-5.5"}
-        usage.write_pending_snapshot(flush_request_id="after-error")
-        assert_pending(
-            pending_path,
-            flows=0,
-            buffered=0,
-            reports=0,
-            flush_request_id="after-error",
-        )
-
     def test_full_pipeline_model_websocket_zero_frame_preserves_billed_usage_and_id(
         self, tmp_path, real_flow
     ):
@@ -1177,186 +855,3 @@ class TestModelProviderWebSocketUsage:
         assert webhook.request_count == 0
         assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
         assert model_provider_usage_sources(flow) == {}
-
-    def test_model_websocket_deferred_trim_keeps_latest_server_message(
-        self,
-        tmp_path,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        old_client = append_websocket_message(flow, from_client=True, content=b"client-old")
-        old_server = append_websocket_message(flow, from_client=False, content=b"server-old")
-        latest_server = append_websocket_message(
-            flow,
-            from_client=False,
-            content=openai_websocket_usage_frame("resp_ws_latest"),
-        )
-        assert flow.websocket is not None
-        messages = flow.websocket.messages
-
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert messages == [old_client, old_server, latest_server]
-        assert model_provider_usage_sources(flow) == {}
-        assert _sum_quantities_by_category(webhook.usage_events()) == {
-            "tokens.input": 10,
-            "tokens.output": 4,
-        }
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert len(deferred_websocket_trim_scheduler) == 1
-
-        run_deferred_websocket_trims(deferred_websocket_trim_scheduler)
-
-        assert flow.websocket.messages is messages
-        assert flow.websocket.messages == [latest_server]
-
-    def test_model_websocket_deferred_trim_keeps_latest_client_message(
-        self,
-        tmp_path,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        old_server = append_websocket_message(flow, from_client=False, content=b"server-old")
-        latest_client = append_websocket_message(
-            flow,
-            from_client=True,
-            content=openai_websocket_usage_frame("resp_ws_client"),
-        )
-
-        mitm_addon.websocket_message(flow)
-
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert model_provider_usage_sources(flow) == {}
-        assert len(deferred_websocket_trim_scheduler) == 1
-
-        run_deferred_websocket_trims(deferred_websocket_trim_scheduler)
-
-        assert flow.websocket is not None
-        assert flow.websocket.messages == [latest_client]
-        assert old_server not in flow.websocket.messages
-
-    def test_model_websocket_deferred_trim_coalesces_and_keeps_latest_at_callback(
-        self,
-        tmp_path,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        first_server = append_websocket_message(
-            flow,
-            from_client=False,
-            content=openai_websocket_usage_frame(
-                "resp_ws_first",
-                input_tokens=1,
-                output_tokens=1,
-            ),
-        )
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            assert len(deferred_websocket_trim_scheduler) == 1
-
-            latest_server = append_websocket_message(
-                flow,
-                from_client=False,
-                content=openai_websocket_usage_frame("resp_ws_latest"),
-            )
-            mitm_addon.websocket_message(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert len(deferred_websocket_trim_scheduler) == 1
-        assert flow.websocket is not None
-        assert flow.websocket.messages == [first_server, latest_server]
-
-        run_deferred_websocket_trims(deferred_websocket_trim_scheduler)
-
-        assert flow.websocket.messages == [latest_server]
-        assert model_provider_usage_sources(flow) == {}
-        assert _sum_quantities_by_category(webhook.usage_events()) == {
-            "tokens.input": 11,
-            "tokens.output": 5,
-        }
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-
-    def test_non_model_websocket_message_retention_is_unchanged(
-        self,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = real_flow(with_response=False, host="example.com")
-        flow.metadata[metadata_keys.VM_RUN_ID] = "run-abc-123"
-        first = append_websocket_message(flow, from_client=True, content=b"client")
-        second = append_websocket_message(flow, from_client=False, content=b"server")
-
-        mitm_addon.websocket_message(flow)
-
-        assert deferred_websocket_trim_scheduler == []
-        assert flow.websocket is not None
-        assert flow.websocket.messages == [first, second]
-
-    def test_model_websocket_end_clears_final_retained_message(
-        self,
-        tmp_path,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        append_websocket_message(
-            flow,
-            from_client=False,
-            content=openai_websocket_usage_frame("resp_ws_1"),
-        )
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            assert len(deferred_websocket_trim_scheduler) == 1
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input": 10,
-            "tokens.output": 4,
-        }
-        assert flow.websocket is not None
-        assert flow.websocket.messages == []
-        assert "model_websocket_usage_enabled" not in flow.metadata
-
-        run_deferred_websocket_trims(deferred_websocket_trim_scheduler)
-        assert flow.websocket.messages == []
-
-    def test_model_websocket_error_clears_final_retained_message(
-        self,
-        tmp_path,
-        real_flow,
-        deferred_websocket_trim_scheduler: list[ScheduledWebSocketTrim],
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        flow.error = Error("connection reset by peer")
-        append_websocket_message(
-            flow,
-            from_client=False,
-            content=openai_websocket_usage_frame("resp_ws_1"),
-        )
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            assert len(deferred_websocket_trim_scheduler) == 1
-            mitm_addon.error(flow)
-            usage.flush_usage_events(trigger="test")
-
-        assert {event["category"]: event["quantity"] for event in webhook.usage_events()} == {
-            "tokens.input": 10,
-            "tokens.output": 4,
-        }
-        assert flow.websocket is not None
-        assert flow.websocket.messages == []
-        assert "model_websocket_usage_enabled" not in flow.metadata
-
-        run_deferred_websocket_trims(deferred_websocket_trim_scheduler)
-        assert flow.websocket.messages == []
